audio_transcription_project/python_backend/requirements.txt
pyserial==3.5
SpeechRecognition==3.10.0
fastapi
uvicorn
websockets
google-genai
python-dotenv
